refactor: 폴더 레벨 변경
